Divide main js code into smaller files
