+++ core/src/navigation/TopNav.html
@@ -87,7 +87,7 @@
   import ContextSwitcher from './ContextSwitcher.html';
   import LogoTitle from './LogoTitle.html';
   import Authorization from '../Authorization.html';
-  import { handleRouteClick } from '../services/routing.js';
+  import { handleRouteClick } from '../services/routing-handlers';
   import * as Navigation from './services/navigation.js';
   import { LuigiConfig } from '../services/config.js';
   import { getNegatedBoolString } from '../utilities/helpers/helpers.js';
